Introduce locks around sections which should run sequentially
Locks are also put on sections which do not modify the exports file,
but mounts the root nfs export. This is a precaution to keep the number
of simultaneous mounts of the root to a minimum, as the NFS connector
has stability issues related to keeping an export mounted locally.